+++ src/shapes/Path.ts
@@ -103,7 +103,16 @@ export class Path extends Shape<PathConfig> {
   getSelfRect() {
     var points = [];
     this.dataArray.forEach(function(data) {
-      points = points.concat(data.points);
+      if (data.command === 'A') {
+        points = points.concat([
+          data.points[0] - data.points[2],
+          data.points[1] - data.points[3],
+          data.points[0] + data.points[2],
+          data.points[1] + data.points[3]
+        ])
+      } else {
+        points = points.concat(data.points);
+      }
     });
     var minX = points[0];
     var maxX = points[0];